Refactor base Action and Filter classes to remove duplication.
Both the Action and Filter class defined a common set of public attributes
along with a default implementation for get_permission and validate.

This common functionality is extracted up to the Element class, which
is a common base class for both Action and Filter.
